+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 ## master (unreleased)
 
+## v2.8.2
+
+### New features
+
+* [#4002](https://github.com/xmake-io/xmake/issues/4002): Add soname and version support
+* [#1613](https://github.com/xmake-io/xmake/issues/1613): Add avx512 and sse4.2 for add_vectorexts
+* [#2471](https://github.com/xmake-io/xmake/issues/2471): Add set_encodings to set source/target encodings
+* [#4071](https://github.com/xmake-io/xmake/pull/4071): Support the stm8 assembler on the sdcc toolchain.
+* [#4101](https://github.com/xmake-io/xmake/issues/4101): Add force includes for c/c++
+* [#2384](https://github.com/xmake-io/xmake/issues/2384): Add extrafiles for vs/vsxmake generator
+
+### Changes
+
+* [#3960](https://github.com/xmake-io/xmake/issues/3960): Improve msys2/crt64 support
+* [#4032](https://github.com/xmake-io/xmake/pull/4032): Remove some old deprecated apis
+* Improve to upgrade vcproj files in tools.msbuild
+* Support add_requires("xmake::xxx") package
+* [#4049](https://github.com/xmake-io/xmake/issues/4049): Improve rust to support cross-compilation
+* Improve clang modules support
+
+### Bugs fixed
+
+* Fix exit all child processes on macOS/Linux
+
 ## v2.8.1
 
 ### New features
@@ -1439,7 +1463,7 @@
 ### Bugs fixed
 
 * [#41](https://github.com/waruqi/xmake/issues/41): Fix checker bug for windows
-* [#43](https://github.com/waruqi/xmake/issues/43): Avoid to generate unnecessary .xmake directory
+* [#43](https://github.com/waruqi/xmake/issues/43): Avoid generating unnecessary .xmake directory
 * Add c++ stl search directories for android
 * Fix compile error for rhel 5.10
 * Fix `os.iorun` bug
@@ -1607,6 +1631,30 @@
 # 更新日志
 
 ## master (开发中)
+
+## v2.8.2
+
+### 新特性
+
+* [#4002](https://github.com/xmake-io/xmake/issues/4002): 增加 soname 支持
+* [#1613](https://github.com/xmake-io/xmake/issues/1613): 为 add_vectorexts 增加 avx512 和 sse4.2 支持
+* [#2471](https://github.com/xmake-io/xmake/issues/2471): 添加 set_encodings API 去设置源文件和目标文件的编码
+* [#4071](https://github.com/xmake-io/xmake/pull/4071): 支持 sdcc 的 stm8 汇编器
+* [#4101](https://github.com/xmake-io/xmake/issues/4101): 为 c/c++ 添加 force includes
+* [#2384](https://github.com/xmake-io/xmake/issues/2384): 为 vs/vsxmake 生成器添加 add_extrafiles 接口
+
+### 改进
+
+* [#3960](https://github.com/xmake-io/xmake/issues/3960): 改进 msys2/crt64 支持
+* [#4032](https://github.com/xmake-io/xmake/pull/4032): 移除一些非常老的废弃接口
+* 改进 tools.msbuild 升级 vcproj 文件
+* 支持 add_requires("xmake::xxx") 包
+* [#4049](https://github.com/xmake-io/xmake/issues/4049): 改进 Rust 支持交叉编译
+* 改进 clang 下 c++ modules 支持
+
+### Bugs 修复
+
+* 修复 macOS/Linux 上子子进程无法快速退出问题
 
 ## v2.8.1